fix: corrige variação do dólar trabalhando com BigDecimal

+++ src/main/java/com/andersonalexdurante/dto/DollarVariationDTO.java
@@ -1,4 +1,6 @@
 package com.andersonalexdurante.dto;
 
-public record DollarVariationDTO(double variation, boolean isUp) {
+import java.math.BigDecimal;
+
+public record DollarVariationDTO(BigDecimal variation, boolean isUp) {
 }
